Share more basic utils
